Adds an edit command
An edit command is badly needed to reduce the length of time it takes a user to update one of their notes. I feel this implementation is a happy tie over until a better solution comes along. There are some drawbacks with this method, mainly being that the note is "recreated" every time it is updated, and therefore has no revision system. This is something to keep in mind when using it.

Signed-off-by: DistroByte <jamesthackett1@gmail.com>